feature/fix: return shuffled input array if numReq === itemCount
Previously, if the `$numReq` parameter's value was equal to the count of the array items, it'd return the original array as-is, with no randomization whatsoever, conflicting with the function's original intentions.
This change aims to fix this behavior by shuffling the original array and returning it if the count of the array items equal to the `$numReq` parameter.

Additionally, this change fixes a possible ValueError by clamping the `$numReq` value to a minimum value of `1` and a maximum value of `$itemCount`.
The error could happen if the end-user would pick either
- a lower `$numReq` than 1 or
- a higher `$numReq` than `$itemCount`